changing approach to use default config
