+++ static/app/views/settings/incidentRules/constants.tsx
@@ -44,22 +44,30 @@ export const errorFieldConfig: OptionConfig = {
   fields: ['user'],
 };
 
+const commonAggregations: AggregationKey[] = [
+  'avg',
+  'percentile',
+  'p50',
+  'p75',
+  'p95',
+  'p99',
+  'p100',
+];
+
+/**
+ * Allowed aggregations for alerts created from wizard
+ */
+export const wizardAlertFieldConfig: OptionConfig = {
+  aggregations: commonAggregations,
+  fields: ['transaction.duration'],
+  measurementKeys: Object.keys(WEB_VITAL_DETAILS),
+};
+
 /**
  * Allowed transaction aggregations for alerts
  */
 export const transactionFieldConfig: OptionConfig = {
-  aggregations: [
-    'avg',
-    'percentile',
-    'failure_rate',
-    'apdex',
-    'count',
-    'p50',
-    'p75',
-    'p95',
-    'p99',
-    'p100',
-  ],
+  aggregations: [...commonAggregations, 'failure_rate', 'apdex', 'count'],
   fields: ['transaction.duration'],
   measurementKeys: Object.keys(WEB_VITAL_DETAILS),
 };
